refactor(onboarding): extract session-reconcile identity resolver from index.ts
The onboarding wiring (identity resolver, submit deps, email sender) grew the
composition root from 778 to 802 lines, tripping the 800-line lint:size cap
(CI-gated). Extract the §5.1 reconcile identity-resolver into
composition/session-reconcile.ts (mirrors resolveEmailSender/resolveGeocoder),
bringing index.ts to 793. Behavior-preserving: 23 reconcile/auth tests green.
