+++ packages/api/src/index.ts
@@ -15,6 +15,7 @@ import {
   resolveOperatorAlertEmail,
   resolvePaymentGateway,
 } from './composition/services'
+import { buildReconcileIdentityResolver } from './composition/session-reconcile'
 import { setupGlobalHandlers } from './error-handlers'
 import { parseBoolFlag } from './lib/parse-bool-flag'
 import {
@@ -378,19 +379,9 @@ export function createApp(overrides?: AppOverrides, repos: Repos = buildRepos(ov
   )
 
   // §5.1 session reconcile: expose the DB-authoritative identity so GET /auth/session
-  // can re-mint a promoted (RENTER->OWNER) token without re-login. Reuses the users
-  // projection (findByIds selects operatorId, drizzle/user.ts) — one indexed read,
-  // and only for non-operator tokens (the handler gates on !isOperatorRole).
-  app.use(
-    '*',
-    provideIdentityResolver(async (user) => {
-      const projection = (await userRepo.findByIds([user.id]))[0]
-      if (!projection) return undefined
-      return projection.operatorId != null
-        ? { role: projection.role, operatorId: projection.operatorId }
-        : { role: projection.role }
-    }),
-  )
+  // can re-mint a promoted (RENTER->OWNER) token without re-login (resolver in
+  // composition/session-reconcile.ts).
+  app.use('*', provideIdentityResolver(buildReconcileIdentityResolver(userRepo)))
 
   // Auth middleware on all protected paths.
   // vehicle-classes: public GETs for renter catalog (list, by-slug, availability)
